ANALYTICS_MODULE/requirements.txt
Django == 2.1
djangorestframework == 3.10.2
psycopg2 == 2.7.7
celery == 3.1.24
billiard == 3.3.0.23
amqp == 1.4.9
numpy == 1.17.0
pandas == 0.25.0
scipy == 1.3.0
scikit-learn == 0.21.3
lightgbm == 2.2.3
joblib == 0.13.2
psutil == 5.6.3
netifaces == 0.10.9
matplotlib == 3.1.1